[#92] Fix runtests scripts proceeding on compilation error (#101)
Runtests scripts compile the program under test, before executing the
tests. The tests always execute regardless of whether the compilation
succeeds or not.

If the program has compilation errors, the runtests scripts are actually
testing an older version of the program without the errors. This results
in misleading test results.

Let's teach the scripts to terminate immediately when the program fails
to compile.