[#103] LearningOutcomes.adoc: fix link to download code (#104)
Before fb74fba, learning outcomes were written in README.adoc.
Developers were instructed to use the 'Clone or download' button
**above** the page, when viewing README.adoc on GitHub.

This instruction no longer make sense after learning outcomes have a
.adoc page on its own. Furthermore, the 'Clone or download' button will
not be available on the rendered static website version anyway, even if
it stayed in README.adoc.

Let's fix the instruction's link to download code, by using the
`site-githuburl` attribute declared in `build.gradle`. That attribute
will not be declared when viewing the .adoc page on GitHub, so declare a
separate attribute for `env-github` environment.