Add netlify.toml
We use Netlify to build and host deploy previews of our documentation
for our PRs. Currently, Netlify is configured to build our documentation
using the command "./gradlew asciidoctor".

Without a netlify.toml file in our repo, only the Nelify account owner
can change that command, which is inconvenient.

Let's add a netlify.toml file to the repo, so that any committer to the
repo can change the Netlify configuration.