runtests.bat: fix filepath in `javac` command
The `javac` command in the `runtests.bat` script reports a "file not
found" error if the case sensitive attribute of the `src` folder is
enabled.

`runtests.bat` script will fail to compile the code into the `bin`
folder and exit prematurely. This results in developers being unable
to execute the tests successfully.

Let's fix the filepath in the `javac` command to allow it to compile
successfully and continue executing the remaining tests even when the
case sensitive attribute of the `src` folder is enabled.