Add checkstyle support
Include a modified version of checkstyle rules based on AB-4 without
`suppressions.xml` as there are no JUnit test files to exclude.

Fix checkstyle violations as AB-1 is supposed to be a model for
students to follow.

Modify the buildfile to run checkstyle as part of the default build
process to ensure that future builds will remain compliant.